refactor: adjust story page, move component to `story/normal/index`
